+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to Killendar are documented here.
 
 Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.2] - Unreleased
+## [1.1.2] - 2026-09-02
 
 1.1.2 improves the calendar layout.